bots: Support a scenario suffix for external project tests
This is useful for running external tests on the same OS with e. g.
different browsers, backends, configurations, etc. We don't want to
encode a set of environment variables into our test names, as that would
get too unwieldy, and we don't have a way to show this in GitHub PRs in
a structured way. So just introduce an abstract "scenario name" that
gets passed to tests in `$TEST_SCENARIO`, that they then can use to map
to corresponding test scenarios or configurations.

Scenarios are specified as a third component of an (external) test name,
such as "cockpit/fedora-28/firefox". If there is no third component,
behavior is unchanged.

Closes #10194