doc: Document proper FreeIPA sudo setup
FreeIPA server does not set up a proper sudo rule by
default (https://pagure.io/freeipa/issue/7538). This is
an "unbreak my setup" setup, but FreeIPA has decided that this is
intended. So instead document the necessary steps as part of the server
requirements. This reflects the `addSudoRule()` hack in check-realms.

Note that it is not appropriate for Cockpit to silently do this by
itself. Joining a domain is a client-side operation and should not
change the server configuration. We also don't know about any deliberate
policy changes on existing servers, so this would only be appropriate as
part of the Cockpit FreeIPA installer UI, not the client joining.

Fixes #10169
Closes #10190