ssh: Test using ~/.ssh/known_hosts
Add a unit test for the user's known_hosts file, which is an important
real-life use case.

Fix set_knownhosts_file() to always check the user's known_hosts file.
This was previously guarded by the system known_hosts file being the
default one; but this was just to make unit tests robust, not because
it's conceptually the right thing. Instead, change unit tests to run
*all* tests with a temporary `$HOME`, to avoid user ssh config to
influence the test results.

Closes #10217